requirements.txt
torch==1.11.0+cu113
transformers==4.18.0
torchmetrics==0.11.0
tokenizers==0.12.1
pytorch-lightning==1.7.7
omegaconf==2.2.3
pillow==9.3.0